Updated Testing Code: Teensy_ReadTweet Compiles
Teensy_ReadTweet now successfully compiles, so it is assumed that
whatever is typed into the serial monitor (directly from the computer
connected to the Teensyduino) will be communicated to the HUZZAH board
over an I2C connection. This has still not been tested, so it is not
certain that this code is production-ready.